refactor(EffectsErrorHandler): Extract to separate file, rename default handler to describe purpose better.
