+++ modules/effects/src/index.ts
@@ -2,7 +2,8 @@ export { createEffect } from './effect_creator';
 export { EffectConfig } from './models';
 export { Effect } from './effect_decorator';
 export { getEffectsMetadata } from './effects_metadata';
-export { mergeEffects, EffectsErrorHandler } from './effects_resolver';
+export { mergeEffects } from './effects_resolver';
+export { EffectsErrorHandler } from './effects_error_handler';
 export { EffectsMetadata, CreateEffectMetadata } from './models';
 export { Actions, ofType } from './actions';
 export { EffectsModule } from './effects_module';
