+++ modules/effects/src/tokens.ts
@@ -1,5 +1,5 @@
 import { InjectionToken, Type } from '@angular/core';
-import { EffectsErrorHandler } from './effects_resolver';
+import { EffectsErrorHandler } from './effects_error_handler';
 
 export const _ROOT_EFFECTS_GUARD = new InjectionToken<void>(
   '@ngrx/effects Internal Root Guard'
